refactor: split effects_engine.py (1050 LOC) into subpackage per AGENTS.md Rule 14
- Create mcp_video/effects_engine/ subpackage with 5 modules:
  core.py (341) - vignette, chromatic_aberration, scanlines, noise, glow
  layout.py (228) - layout_grid, layout_pip
  text.py (205) - text_animated, text_subtitles
  mograph.py (203) - mograph_count, mograph_progress
  utility.py (164) - video_info_detailed, auto_chapters
- __init__.py re-exports all public functions preserving backward-compatible
  import path (from mcp_video.effects_engine import effect_vignette, ...)
- Update test_effects_engine.py monkeypatch targets to use
  effects_engine.core._run_ffmpeg

Tests: 773 passed, 9 skipped, 2 xpassed
