+++ tests/test_effects_engine.py
@@ -19,7 +19,7 @@ def test_chromatic_aberration_falls_back_when_chromashift_missing(tmp_path, monk
             raise ProcessingError(" ".join(cmd), 1, "No such filter: 'chromashift'")
         Path(cmd[-1]).write_bytes(b"output")
 
-    monkeypatch.setattr(effects_engine, "_run_ffmpeg", fake_run_ffmpeg)
+    monkeypatch.setattr(effects_engine.core, "_run_ffmpeg", fake_run_ffmpeg)
 
     result = effects_engine.effect_chromatic_aberration(str(input_path), str(output_path), intensity=3.0)
 
@@ -43,7 +43,7 @@ def test_glow_falls_back_when_gblur_missing(tmp_path, monkeypatch):
             raise ProcessingError(" ".join(cmd), 1, "No such filter: 'gblur'")
         Path(cmd[-1]).write_bytes(b"output")
 
-    monkeypatch.setattr(effects_engine, "_run_ffmpeg", fake_run_ffmpeg)
+    monkeypatch.setattr(effects_engine.core, "_run_ffmpeg", fake_run_ffmpeg)
 
     result = effects_engine.effect_glow(str(input_path), str(output_path), radius=15)
 
@@ -64,7 +64,7 @@ def test_chromatic_aberration_reraises_unrelated_processing_error(tmp_path, monk
     def fake_run_ffmpeg(cmd):
         raise error
 
-    monkeypatch.setattr(effects_engine, "_run_ffmpeg", fake_run_ffmpeg)
+    monkeypatch.setattr(effects_engine.core, "_run_ffmpeg", fake_run_ffmpeg)
 
     try:
         effects_engine.effect_chromatic_aberration(str(input_path), str(output_path))
